Day 8
I'm a little surprised that took as long as it did -- I spent about 2.5 hours on this.

I didn't think to memoize (which lead to  `compute_distances`) until after spending 90 minutes on it the first night. And then had bugs in `join` for most of the rest of the time. Rusty.